fix: follow a stricter compiler set of rules, brrrrrrrrr
 - replace c style cast to static_cast<>     |
 - remove some #define with static constexpr |
 - use size_t for STL containers loop        |
 - fix other typing conversion               |
 - fix various shadow variable               |
_____________________________________________

  , ; ,   .-'"""'-.   , ; ,
  \\|/  .'         '.  \|//
   \-;-/   ()   ()   \-;-/
   // ;               ; \\
  //__; :.         .; ;__\\
 `-----\'.'-.....-'.'/-----'
        '.'.-.-,_.'.'
          '(  (..-'
            '-'

+++ src/engine/sprite_animation.cpp
@@ -37,7 +37,7 @@ void SpriteAnimation::render(SDL_Renderer *renderer, float parent_x, float paren
         SDL_FLIP_NONE
     );
 
-    int delta = SDL_GetTicks();
+    Uint64 delta = SDL_GetTicks();
 
     if(delta > this->last_update_time + this->frame_update_interval) {
         this->frame_index++;
